Major refactor of simulation functions
