+++ frontend/src/constants/goerli-constants.ts
@@ -1,4 +1,4 @@
-import { ConstantsObject } from "../types/manual/types";
+import { ElementAddresses } from "../types/manual/types";
 
 export interface Tranche {
     expiration: number,
@@ -19,7 +19,7 @@ export interface Tranche {
     convergentCurvePoolFactory: string,
 }
 
-export const constants: ConstantsObject = {
+export const constants: ElementAddresses = {
 	tokens: {
 		usdc: "0x78dEca24CBa286C0f8d56370f5406B48cFCE2f86",
 		weth: "0x9A1000D492d40bfccbc03f413A48F5B6516Ec0Fd",
