+++ frontend/src/constants/mainnet-constants.ts
@@ -1,6 +1,6 @@
-import { ConstantsObject } from "../types/manual/types";
+import { ElementAddresses } from "../types/manual/types";
 
-export const constants: ConstantsObject = {
+export const constants: ElementAddresses = {
 	"tokens": {
 		"usdc": "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
 		"weth": "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
